avoid a variety of warnings for MSVC
git-svn-id: https://svn.osgeo.org/gdal/trunk@21684 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/ogr/ogrsf_frmts/dgn/dgnread.cpp
@@ -926,7 +926,7 @@ static DGNElemCore *DGNProcessElement( DGNInfo *psDGN, int nType, int nLevel )
       case DGNT_BSPLINE_SURFACE_BOUNDARY:
         {
           DGNElemBSplineSurfaceBoundary *psBounds;
-          int numverts = psDGN->abyElem[38] + psDGN->abyElem[39]*256;
+          short numverts = psDGN->abyElem[38] + psDGN->abyElem[39]*256;
 
           psBounds = (DGNElemBSplineSurfaceBoundary *)
             CPLCalloc(sizeof(DGNElemBSplineSurfaceBoundary)+
